Add verification API facade
