+++ dev-requirements.txt
@@ -1,5 +1,5 @@
 #
-# This file is autogenerated by pip-compile with Python 3.11
+# This file is autogenerated by pip-compile with Python 3.10
 # by the following command:
 #
 #    pip-compile dev-requirements.in
@@ -18,6 +18,8 @@ distlib==0.3.8
     # via virtualenv
 docopt==0.6.2
     # via tbump
+exceptiongroup==1.3.0
+    # via pytest
 filelock==3.16.1
     # via virtualenv
 freezegun==1.5.1
@@ -60,8 +62,15 @@ tabulate==0.8.10
     # via cli-ui
 tbump==6.11.0
     # via -r dev-requirements.in
+tomli==2.2.1
+    # via
+    #   build
+    #   pip-tools
+    #   pytest
 tomlkit==0.11.8
     # via tbump
+typing-extensions==4.14.0
+    # via exceptiongroup
 unidecode==1.3.8
     # via cli-ui
 virtualenv==20.26.6
