+++ soda-core/src/soda_core/contracts/__init__.py
@@ -1,0 +1,11 @@
+from soda_core.contracts.api import (
+    verify_contracts,
+    verify_contracts_locally,
+    verify_contracts_on_agent,
+)
+
+__all__ = [
+    "verify_contracts",
+    "verify_contracts_locally",
+    "verify_contracts_on_agent",
+]
